Move commonly-used swap fixture to conftest.
