+++ tests/conftest.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import Session
 from app.dependencies.database import get_session
 from app.dependencies.notifications import Notification, get_notification_service
 from app.main import app
-from app.models import Base
+from app.models import Base, Game, Gamer, Swap
 
 
 class NotificationServiceMock:
@@ -42,4 +42,26 @@ def client_fixture(session: Session) -> Generator[TestClient, None, None]:
     client = TestClient(app)  
     yield client  
     app.dependency_overrides.clear()  
+
+
+@pytest.fixture
+def swap(session: Session) -> Swap:
+    proposer = Gamer(name="Player One", email="press@start.com")
+    acceptor = Gamer(name="Player Two", email="insert@coin.com")
+    session.add_all([proposer, acceptor])
+    session.commit()
+
+    proposer_game = Game(title="Sonic The Hedgehog", platform="SEGA Mega Drive", gamer_id=proposer.id)
+    acceptor_game = Game(title="Super Mario Land", platform="Nintendo GAME BOY", gamer_id=acceptor.id)
+    session.add_all([proposer_game, acceptor_game])
+    session.commit()
+
+    swap = Swap(proposer_id=proposer.id, acceptor_id=acceptor.id)
+    session.add(swap)
+    session.commit()
+    swap.games.append(proposer_game)
+    swap.games.append(acceptor_game)
+    session.commit()
+
+    return swap
