+++ tests/test_swaps.py
@@ -1,31 +1,7 @@
-import pytest
-
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
 from app.models import Game, Gamer, Swap
-
-
-@pytest.fixture
-def swap(session: Session) -> Swap:
-    proposer = Gamer(name="Player One", email="press@start.com")
-    acceptor = Gamer(name="Player Two", email="insert@coin.com")
-    session.add_all([proposer, acceptor])
-    session.commit()
-
-    proposer_game = Game(title="Sonic The Hedgehog", platform="SEGA Mega Drive", gamer_id=proposer.id)
-    acceptor_game = Game(title="Super Mario Land", platform="Nintendo GAME BOY", gamer_id=acceptor.id)
-    session.add_all([proposer_game, acceptor_game])
-    session.commit()
-
-    swap = Swap(proposer_id=proposer.id, acceptor_id=acceptor.id)
-    session.add(swap)
-    session.commit()
-    swap.games.append(proposer_game)
-    swap.games.append(acceptor_game)
-    session.commit()
-
-    return swap
 
 
 def test_create_swap(session: Session, client: TestClient) -> None:
